requirements.txt
asgiref==3.8.1
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
